Next we can build the controls. The user will be able to directly select the track by clicking on it, so there is no need for any additional interface there. We will eventually add a volume control, but let's not jump ahead until we have some music playing.
We can use the same button for play or pause and simply change the icon on it depending on the status. Vue.js lets us easily apply classes according to either a variable or a function. This is a feature I have used heavily in small, interactive Vue.js applications in order to denote some sort of state. Here we simply need to set the class based on whether the music is playing or paused. We could be direct here and write `'fa-play-circle': STATUSES.PAUSED === status`. It's often better to take these seemingly simple equality checks and move them to a computed property. That not only makes it clearer what we're checking, but it makes later updates easier.

We know that some action will need to happen when the user toggles this button. Since we don't know what that action is yet, we'll add a placeholder method in our application.