Everything is in place now to play some music - we have our track listing, and our play/pause button. Thankfully, browsers have made it exceptionally easy to play audio. We're going to control the music using the <a href='https://developer.mozilla.org/en-US/docs/Web/API/HTMLAudioElement' target='_blank'>HTMLAudioElement</a> interface. It allows us to create an audio element by simply calling `new Audio([URLString])`, and then simply call `play` and `pause` functions to control playback.
We already have a placeholder for our play/pause button, it's time to think through the logic. First, it needs to ensure that a track is loaded. We'll defer this to their own computed property and method since it doesn't belong in our toggle function. At some point we have to load the track into an `Audio` element, which we can add to our component's data properties. Then it needs to determine if the music should be played or paused by checking the current status. Finally, we will add play and pause methods of our own, which will allow us to toggle the status or add in other functionality later.

<blockquote><em>By default I was loading Font Awesome's icons using their JavaScript method. This made changing the icon for the play/pause button dynamically more difficult, so I switched to using their CSS file instead. Much simpler.</em></blockquote>