Great! A user can now play and pause the first song from our mix tape! Next, we want to indicate which song is playing, and allow the user to play a different song by clicking on it. Because we have been thinking through how we build the application and separating concerns into their own functions, most of this functionality already exists.
Our app tracks the active song by its index, all we need to do is apply a CSS class to the list element if it matches the active song's index. We can get the index of a `for` loop with Vue.js by doing `v-for="(item, index) in list"`.

Loading a track by clicking on it is as simple as associating a click event with the `loadTrack` method, once again using the index value.