Look at that! We have a simple but functional media player. And it even looks pretty good. There are two features from our requirements that have not yet been met - autoplaying the next song when the current one finishes, and controlling the volume.
Autoplaying the next song is relatively straightforward. The `Audio` element provides an `ended` event whenever playback finishes. Our application needs to listen for that, then load the next track if it exists. We can set that up when we create the new `Audio` element. Instead of modifying our `loadTrack` method to play the next track, we'll pull that out into its own method.