Now we need a volume control. It seems that every different audio and video player creates their own variation on this controller, but we're going to go with what HTML5 provides us and give it a bit of CSS polish.
<blockquote><em>While the `Audio` element's volume range is from 0.0 to 1.0, Chrome was giving me fits when I set the value of the range slider to this scale. It would always peg the default value high. I simply changed the scale from 0 to 10 on the slider and scale it down in the code.</em></blockquote>

While the Vue.js docs warns about overusing `watch`, I think this is a good use case for it. We have our global `volume` property that the user will be changing using the range slider. We then need to ensure the current `Audio` element gets updated with the new value. We can do this by setting a `watch` function on the `volume` property.

Let's also not forget to set the volume of a new `Audio` element any time we load a track.