That's it! We have now met all of our requirements. I can think of about 40 more features to add to this, but there is only 1 more I want to deal with for this tutorial. That's a playback progress bar. We need two pieces of information for this - the duration of the song, and the current time.
Getting the song's duration would be as easy as checking the `duration` property on the `Audio` element, except that when initially created the element doesn't yet know the duration. It's not until it downloads the track's metadata that it has this information. We can hook into the `durationchanged` event to update our app's `duration` value as soon as the metadata is loaded.

We'll also add a computed property that converts the duration from the default seconds value to a human-friendly format.

Similarly, we can add an event listener to the element's `timeupdate` event to track the playback progress.

<blockquote><em>I ran into difficulties with this event. When using `addEventListener` my listener would not get called. Setting the `ontimeupdate` seemed to work, so that's what I went with in the code. I'm not sure if this is a bug with the browsers or if I'm missing something really obvious.</em></blockquote>

The easiest way to approach this is to create an `updateProgress` method for the event listener. That will be called every time the song's playback time changes. It then updates the `progress` property we added to our application while converting it to a percentage. This percentage is used to determine the width of a progress bar we add underneath our player's controls. For good measure, we show the total duration of the song.