<h3>Improving our Vue.js Media Player</h3>
There are many improvements to be made to this very basic player. It's great to follow along with tutorials, but nothing can take the place of writing your own code. Consider taking a couple of these features and building them out on top of the base player.

<ul>
  <li>allow the user to click the progress bar to change the current time of the song</li>
  <li>set a cookie to remember the user's volume choice between page loads</li>
  <li>allow the user to put a song on repeat, or play the songs in a random order</li>
</ul>

Send me your results! mwhager87 at gmail.